refactor: consolidate miniapp and modularize edge
