--- miniapp/postcss.config.js
@@ -1,3 +1,0 @@
-export default {
-  plugins: {},
-};
