--- miniapp/src/style.css
@@ -1,159 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-body {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
-  margin: 0;
-  padding: 1rem;
-  background: var(--tg-theme-bg-color, #ffffff);
-  color: var(--tg-theme-text-color, #000000);
-  min-height: 100vh;
-}
-
-#app {
-  max-width: 400px;
-  margin: 0 auto;
-}
-
-.card {
-  background: var(--tg-theme-secondary-bg-color, #f1f3f4);
-  border-radius: 12px;
-  padding: 1.5rem;
-  margin-bottom: 1rem;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.1);
-}
-
-.coming-soon {
-  text-align: center;
-}
-
-h1 {
-  margin: 0 0 0.5rem 0;
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: var(--tg-theme-text-color, #000000);
-}
-
-h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: var(--tg-theme-text-color, #000000);
-}
-
-.muted {
-  color: var(--tg-theme-hint-color, #708499);
-  font-size: 0.9rem;
-  line-height: 1.4;
-}
-
-.row {
-  display: flex;
-  gap: 0.5rem;
-  margin: 1rem 0;
-}
-
-.btn {
-  background: var(--tg-theme-button-color, #007aff);
-  color: var(--tg-theme-button-text-color, #ffffff);
-  border: none;
-  border-radius: 8px;
-  padding: 0.75rem 1rem;
-  font-size: 0.9rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: opacity 0.2s;
-  flex: 1;
-}
-
-.btn:hover {
-  opacity: 0.8;
-}
-
-.btn.secondary {
-  background: var(--tg-theme-secondary-bg-color, #f1f3f4);
-  color: var(--tg-theme-text-color, #000000);
-  border: 1px solid var(--tg-theme-hint-color, #708499);
-}
-
-.btn:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.kv {
-  display: flex;
-  justify-content: space-between;
-  padding: 0.5rem 0;
-  border-bottom: 1px solid var(--tg-theme-hint-color, #e5e5ea);
-}
-
-.kv:last-child {
-  border-bottom: none;
-}
-
-code {
-  background: var(--tg-theme-secondary-bg-color, #f1f3f4);
-  padding: 0.25rem 0.5rem;
-  border-radius: 4px;
-  font-family: 'SF Mono', Monaco, 'Cascadia Code', 'Roboto Mono', Consolas, 'Courier New', monospace;
-  font-size: 0.8rem;
-  word-break: break-all;
-}
-
-.deposit-form {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.input-group {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-label {
-  font-weight: 500;
-  color: var(--tg-theme-text-color, #000000);
-}
-
-input {
-  background: var(--tg-theme-bg-color, #ffffff);
-  border: 1px solid var(--tg-theme-hint-color, #c7c7cc);
-  border-radius: 8px;
-  padding: 0.75rem;
-  font-size: 1rem;
-  color: var(--tg-theme-text-color, #000000);
-}
-
-input:focus {
-  outline: none;
-  border-color: var(--tg-theme-button-color, #007aff);
-}
-
-.status {
-  padding: 0.75rem;
-  border-radius: 8px;
-  text-align: center;
-  font-weight: 500;
-}
-
-.status.success {
-  background: #d4edda;
-  color: #155724;
-  border: 1px solid #c3e6cb;
-}
-
-.status.error {
-  background: #f8d7da;
-  color: #721c24;
-  border: 1px solid #f5c6cb;
-}
-
-.status.processing {
-  background: #d1ecf1;
-  color: #0c5460;
-  border: 1px solid #bee5eb;
-}
