--- miniapp/vite.config.ts
@@ -1,19 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react-swc';
-
-export default defineConfig({
-  plugins: [react()],
-  // Serve built assets from /miniapp/assets when deployed under /miniapp
-  base: '/miniapp/assets/',
-  build: {
-    outDir: 'dist',
-    assetsDir: '',
-    rollupOptions: {
-      output: {
-        entryFileNames: 'index-[hash].js',
-        chunkFileNames: '[name]-[hash].js',
-        assetFileNames: 'index-[hash].[ext]'
-      }
-    }
-  }
-});
