+++ package.json
@@ -11,7 +11,7 @@
     "test": "deno test --no-npm --node-modules-dir=false --unsafely-ignore-certificate-errors=deno.land,cdn.skypack.dev -A --no-check",
     "preview": "vite preview",
     "setup:supabase": "bash scripts/setup-supabase-cli.sh",
-    "build:miniapp": "cd miniapp && npm ci && npm run build && rm -rf ../supabase/functions/miniapp/static && mkdir -p ../supabase/functions/miniapp/static && cp -r dist/* ../supabase/functions/miniapp/static",
+    "build:miniapp": "deno run -A scripts/build-miniapp.ts",
     "deploy:bot": "npx supabase functions deploy telegram-bot --project-ref $PROJECT_REF",
     "deploy:edge": "npx supabase functions deploy miniapp telegram-bot receipt-submit receipt-ocr crypto-webhook admin-act-on-payment --project-ref qeejuomcapbdlhnjqjcc",
     "edge:deploy:core": "npx supabase functions deploy telegram-bot miniapp --project-ref qeejuomcapbdlhnjqjcc",
